Increase tado API polling interval to 5 minutes
Polling interval of 20 seconds causes high load on tado servers and does
not provide enough value to warrant it. tado plans to introduce a rate
limit to prevent such misuse of the API, therefore the polling interval
needs to be increased to make sure the integration works well in the
future.

Weather polling interval was increased to 15 minutes as that is the
granularity on tado side anyway.